auto_commit

+++ upload_huggingface_calculator.py
@@ -10,15 +10,17 @@ OUTPUT_DIR = "mnt/output/qwen3.5-0.8b-lora-calculator"
 REPO_ID = "khanh2023/qwen3.5-0.8b-lora-calculator"
 CODE_SRC = "leanlm"
 
-CODE_DST = f"{OUTPUT_DIR}/src/{CODE_SRC}"
 
-if os.path.exists(CODE_DST):
-    shutil.rmtree(CODE_DST)
+if os.path.exists(CODE_SRC):
+    CODE_DST = f"{OUTPUT_DIR}/src/{CODE_SRC}"
 
-if not os.path.exists(os.path.dirname(CODE_DST)):
-    os.makedirs(os.path.dirname(CODE_DST))
+    if os.path.exists(CODE_DST):
+        shutil.rmtree(CODE_DST)
 
-shutil.copytree(CODE_SRC, CODE_DST)
+    if not os.path.exists(os.path.dirname(CODE_DST)):
+        os.makedirs(os.path.dirname(CODE_DST))
+
+    shutil.copytree(CODE_SRC, CODE_DST)
 
 login()
 upload_large_folder(
